+++ Changelog.md
@@ -17,5 +17,6 @@ YogaSMC Changelog
 - NC: override unknown events with updated presets
 - BaseService: support extended EC operation
 - YogaSMC: add constants for fan support
+- ThinkSMC: support fan speed reading
 
 Change of earlier versions may be tracked in [commits](https://github.com/zhen-zen/YogaSMC/commits/master) and [releases](https://github.com/zhen-zen/YogaSMC/releases).